also fix up stores tests
